More frontend organizing
