+++ client/src/components/graph/nodes/AverageNode.tsx
@@ -8,7 +8,10 @@ import {Handle, NodeProps, Position, useUpdateNodeInternals} from 'reactflow';
 
 import {AverageNodeSettings, AverageNodeValue} from '@/common/types/graph';
 import BaseNode from './BaseNode';
-import {NodeDataContext, NodeValuesContext} from '../../context/GraphProvider';
+import {
+  NodeDataContext,
+  NodeValuesContext,
+} from '../../../context/GraphProvider';
 
 type LocalSettings = {
   weights: {[key: string]: string};
