+++ client/src/components/graph/nodes/CoursePartNode.tsx
@@ -6,18 +6,24 @@ import {ChangeEvent, JSX, useContext, useEffect, useState} from 'react';
 import {useTranslation} from 'react-i18next';
 import {Handle, NodeProps, Position} from 'reactflow';
 
-import {MinPointsNodeSettings, MinPointsNodeValue} from '@/common/types/graph';
+import {
+  CoursePartNodeSettings,
+  CoursePartNodeValue,
+} from '@/common/types/graph';
 import BaseNode from './BaseNode';
-import {NodeDataContext, NodeValuesContext} from '../../context/GraphProvider';
+import {
+  NodeDataContext,
+  NodeValuesContext,
+} from '../../../context/GraphProvider';
 
 type OnFailSetting = 'coursefail' | 'fail';
 type LocalSettings = {onFailSetting: OnFailSetting; minPoints: string};
 const initialSettings: LocalSettings = {
   onFailSetting: 'coursefail',
-  minPoints: '0',
+  minPoints: '',
 };
 
-const MinPointsNode = (props: NodeProps): JSX.Element => {
+const CoursePartNode = (props: NodeProps): JSX.Element => {
   const {t} = useTranslation();
   const {id, isConnectable} = props;
 
@@ -28,15 +34,19 @@ const MinPointsNode = (props: NodeProps): JSX.Element => {
   const [error, setError] = useState<boolean>(false);
   const [init, setInit] = useState<boolean>(false);
 
-  const nodeValue = nodeValues[id] as MinPointsNodeValue;
-  const settings = nodeData[id].settings as MinPointsNodeSettings;
+  const nodeValue = nodeValues[id] as CoursePartNodeValue;
+  const settings = nodeData[id].settings as CoursePartNodeSettings;
 
   useEffect(() => {
     if (init) return;
-    setLocalSettings({...settings, minPoints: settings.minPoints.toString()});
+    setLocalSettings({
+      ...settings,
+      minPoints:
+        settings.minPoints !== null ? settings.minPoints.toString() : '',
+    });
     setError(false);
     setInit(true);
-  }, [nodeData]); // eslint-disable-line react-hooks/exhaustive-deps
+  }, [nodeValues]); // eslint-disable-line react-hooks/exhaustive-deps
 
   const handleSelectChange = (event: ChangeEvent<HTMLSelectElement>): void => {
     if (localSettings.onFailSetting === event.target.value) return;
@@ -60,7 +70,10 @@ const MinPointsNode = (props: NodeProps): JSX.Element => {
     const newLocalSettings = {...localSettings, minPoints: event.target.value};
     setLocalSettings(newLocalSettings);
 
-    if (!/^\d+(?:\.\d+?)?$/.test(newLocalSettings.minPoints)) {
+    if (
+      newLocalSettings.minPoints !== '' &&
+      !/^\d+(?:\.\d+?)?$/.test(newLocalSettings.minPoints)
+    ) {
       setError(true);
       return;
     }
@@ -68,20 +81,15 @@ const MinPointsNode = (props: NodeProps): JSX.Element => {
 
     setNodeSettings(id, {
       ...newLocalSettings,
-      minPoints: parseFloat(newLocalSettings.minPoints),
+      minPoints:
+        newLocalSettings.minPoints === ''
+          ? null
+          : parseFloat(newLocalSettings.minPoints),
     });
   };
 
   return (
     <BaseNode {...props} error={error} courseFail={nodeValue.courseFail}>
-      <Handle
-        type="target"
-        id={id}
-        style={{height: '12px', width: '12px'}}
-        position={Position.Left}
-        isConnectable={isConnectable}
-      />
-
       <div>
         <label>{t('graph.min-points')}: </label>
         <input
@@ -91,16 +99,18 @@ const MinPointsNode = (props: NodeProps): JSX.Element => {
           value={localSettings.minPoints}
         />
       </div>
-      <div style={{textAlign: 'left'}}>
-        <label>{t('graph.on-fail')}: </label>
-        <select
-          onChange={handleSelectChange}
-          value={localSettings.onFailSetting}
-        >
-          <option value="coursefail">{t('graph.fail-course')}</option>
-          <option value="fail">{t('graph.output-fail')}</option>
-        </select>
-      </div>
+      {settings.minPoints !== null && (
+        <div style={{textAlign: 'left'}}>
+          <label>On fail: </label>
+          <select
+            onChange={handleSelectChange}
+            value={localSettings.onFailSetting}
+          >
+            <option value="coursefail">{t('graph.fail-course')}</option>
+            <option value="fail">{t('graph.output-fail')}</option>
+          </select>
+        </div>
+      )}
       <p className="output-value">
         {t('graph.output')}:{' '}
         {nodeValue.value === 'fail'
@@ -119,4 +129,4 @@ const MinPointsNode = (props: NodeProps): JSX.Element => {
   );
 };
 
-export default MinPointsNode;
+export default CoursePartNode;
