+++ client/src/components/graph/nodes/GradeNode.tsx
@@ -8,7 +8,7 @@ import {Handle, NodeProps, Position} from 'reactflow';
 
 import {GradeNodeValue} from '@/common/types/graph';
 import BaseNode from './BaseNode';
-import {NodeValuesContext} from '../../context/GraphProvider';
+import {NodeValuesContext} from '../../../context/GraphProvider';
 
 const GradeNode = (props: NodeProps): JSX.Element => {
   const {t} = useTranslation();
