+++ client/src/components/graph/nodes/MinPointsNode.tsx
@@ -6,21 +6,21 @@ import {ChangeEvent, JSX, useContext, useEffect, useState} from 'react';
 import {useTranslation} from 'react-i18next';
 import {Handle, NodeProps, Position} from 'reactflow';
 
-import {
-  CoursePartNodeSettings,
-  CoursePartNodeValue,
-} from '@/common/types/graph';
+import {MinPointsNodeSettings, MinPointsNodeValue} from '@/common/types/graph';
 import BaseNode from './BaseNode';
-import {NodeDataContext, NodeValuesContext} from '../../context/GraphProvider';
+import {
+  NodeDataContext,
+  NodeValuesContext,
+} from '../../../context/GraphProvider';
 
 type OnFailSetting = 'coursefail' | 'fail';
 type LocalSettings = {onFailSetting: OnFailSetting; minPoints: string};
 const initialSettings: LocalSettings = {
   onFailSetting: 'coursefail',
-  minPoints: '',
+  minPoints: '0',
 };
 
-const CoursePartNode = (props: NodeProps): JSX.Element => {
+const MinPointsNode = (props: NodeProps): JSX.Element => {
   const {t} = useTranslation();
   const {id, isConnectable} = props;
 
@@ -31,19 +31,15 @@ const CoursePartNode = (props: NodeProps): JSX.Element => {
   const [error, setError] = useState<boolean>(false);
   const [init, setInit] = useState<boolean>(false);
 
-  const nodeValue = nodeValues[id] as CoursePartNodeValue;
-  const settings = nodeData[id].settings as CoursePartNodeSettings;
+  const nodeValue = nodeValues[id] as MinPointsNodeValue;
+  const settings = nodeData[id].settings as MinPointsNodeSettings;
 
   useEffect(() => {
     if (init) return;
-    setLocalSettings({
-      ...settings,
-      minPoints:
-        settings.minPoints !== null ? settings.minPoints.toString() : '',
-    });
+    setLocalSettings({...settings, minPoints: settings.minPoints.toString()});
     setError(false);
     setInit(true);
-  }, [nodeValues]); // eslint-disable-line react-hooks/exhaustive-deps
+  }, [nodeData]); // eslint-disable-line react-hooks/exhaustive-deps
 
   const handleSelectChange = (event: ChangeEvent<HTMLSelectElement>): void => {
     if (localSettings.onFailSetting === event.target.value) return;
@@ -67,10 +63,7 @@ const CoursePartNode = (props: NodeProps): JSX.Element => {
     const newLocalSettings = {...localSettings, minPoints: event.target.value};
     setLocalSettings(newLocalSettings);
 
-    if (
-      newLocalSettings.minPoints !== '' &&
-      !/^\d+(?:\.\d+?)?$/.test(newLocalSettings.minPoints)
-    ) {
+    if (!/^\d+(?:\.\d+?)?$/.test(newLocalSettings.minPoints)) {
       setError(true);
       return;
     }
@@ -78,15 +71,20 @@ const CoursePartNode = (props: NodeProps): JSX.Element => {
 
     setNodeSettings(id, {
       ...newLocalSettings,
-      minPoints:
-        newLocalSettings.minPoints === ''
-          ? null
-          : parseFloat(newLocalSettings.minPoints),
+      minPoints: parseFloat(newLocalSettings.minPoints),
     });
   };
 
   return (
     <BaseNode {...props} error={error} courseFail={nodeValue.courseFail}>
+      <Handle
+        type="target"
+        id={id}
+        style={{height: '12px', width: '12px'}}
+        position={Position.Left}
+        isConnectable={isConnectable}
+      />
+
       <div>
         <label>{t('graph.min-points')}: </label>
         <input
@@ -96,18 +94,16 @@ const CoursePartNode = (props: NodeProps): JSX.Element => {
           value={localSettings.minPoints}
         />
       </div>
-      {settings.minPoints !== null && (
-        <div style={{textAlign: 'left'}}>
-          <label>On fail: </label>
-          <select
-            onChange={handleSelectChange}
-            value={localSettings.onFailSetting}
-          >
-            <option value="coursefail">{t('graph.fail-course')}</option>
-            <option value="fail">{t('graph.output-fail')}</option>
-          </select>
-        </div>
-      )}
+      <div style={{textAlign: 'left'}}>
+        <label>{t('graph.on-fail')}: </label>
+        <select
+          onChange={handleSelectChange}
+          value={localSettings.onFailSetting}
+        >
+          <option value="coursefail">{t('graph.fail-course')}</option>
+          <option value="fail">{t('graph.output-fail')}</option>
+        </select>
+      </div>
       <p className="output-value">
         {t('graph.output')}:{' '}
         {nodeValue.value === 'fail'
@@ -126,4 +122,4 @@ const CoursePartNode = (props: NodeProps): JSX.Element => {
   );
 };
 
-export default CoursePartNode;
+export default MinPointsNode;
