+++ client/src/components/graph/nodes/RoundNode.tsx
@@ -8,7 +8,10 @@ import {Handle, NodeProps, Position} from 'reactflow';
 
 import {RoundNodeSettings, RoundNodeValue} from '@/common/types/graph';
 import BaseNode from './BaseNode';
-import {NodeDataContext, NodeValuesContext} from '../../context/GraphProvider';
+import {
+  NodeDataContext,
+  NodeValuesContext,
+} from '../../../context/GraphProvider';
 
 type RoundSetting = 'round-up' | 'round-closest' | 'round-down';
 type LocalSettings = {roundingSetting: RoundSetting};
