+++ client/src/components/graph/nodes/StepperNode.tsx
@@ -8,7 +8,10 @@ import {Handle, NodeProps, Position} from 'reactflow';
 
 import {StepperNodeSettings, StepperNodeValue} from '@/common/types/graph';
 import BaseNode from './BaseNode';
-import {NodeDataContext, NodeValuesContext} from '../../context/GraphProvider';
+import {
+  NodeDataContext,
+  NodeValuesContext,
+} from '../../../context/GraphProvider';
 
 type LocalSettings = {
   numSteps: number;
