+++ client/src/components/graph/nodes/SubstituteNode.tsx
@@ -11,7 +11,10 @@ import {
   SubstituteNodeValue,
 } from '@/common/types/graph';
 import BaseNode from './BaseNode';
-import {NodeDataContext, NodeValuesContext} from '../../context/GraphProvider';
+import {
+  NodeDataContext,
+  NodeValuesContext,
+} from '../../../context/GraphProvider';
 
 type LocalSettings = {maxSubstitutions: string; substituteValues: string[]};
 const initialSettings = {maxSubstitutions: '0', substituteValues: []};
