+++ client/src/tests/UserButton.test.tsx
@@ -14,7 +14,7 @@ import {userEvent} from '@testing-library/user-event';
 import {BrowserRouter} from 'react-router-dom';
 
 import {AuthData, SystemRole} from '@/common/types';
-import UserButton from '../components/auth/UserButton';
+import UserButton from '../components/app-container/UserButton';
 import AuthContext from '../context/AuthProvider';
 
 afterEach(cleanup);
